refactor: move access logic to auth folder + dedicated types file
