+++ src/auth/index.ts
@@ -1,10 +1,6 @@
 import { Page } from 'puppeteer';
 import { FIAP_URLS } from '../constants';
-
-interface LoginCredentials {
-  username: string;
-  password: string;
-}
+import { LoginCredentials } from './types';
 
 export async function loginToFIAP(page: Page, { username, password }: LoginCredentials) {
   await page.goto(FIAP_URLS.LOGIN);
@@ -24,4 +20,17 @@ export async function loginToFIAP(page: Page, { username, password }: LoginCrede
     .catch((e) => {
       console.error('Failed to login:', e);
     });
+}
+
+export async function ensureAccessToCoursePage(page: Page) {
+  await page.goto(FIAP_URLS.COURSE, { waitUntil: 'networkidle0' });
+
+  const courseTitleSelector = '.conteudo-digital-disciplina-unidade';
+  const courseTitle = await page.$(courseTitleSelector);
+
+  if (!courseTitle) {
+    throw new Error('Failed to access course page');
+  }
+
+  console.log('Successfully accessing the subjects page');
 }
