+++ src/auth/types.ts
@@ -1,0 +1,4 @@
+export interface LoginCredentials {
+  username: string;
+  password: string;
+}
